requirements.txt
# langchain==0.3.26
# flask==3.1.1 
# sentence-transformers==4.1.0
# pypdf==5.6.1 
# python-dotenv==1.1.0
# langchain-pinecone==0.2.8 
# langchain-openai==0.3.24
# langchain-community==0.3.26
# -e .

# Flask + environment
flask>=3.1.1
python-dotenv>=1.1.0

# LangChain stack
langchain>=0.3.26
langchain-core>=0.3.0
langchain-community>=0.3.26
langchain-groq>=0.3.0

# Vector store & embeddings
chromadb>=0.6.0
sentence-transformers>=4.1.0
torch>=2.2.0
transformers>=4.40.0

# PDF + other utils
pypdf>=5.6.1
requests>=2.33.1